requirements.txt
Flask==3.0.3
spacy==3.7.5
nltk==3.9.1
scikit-learn==1.5.2
openai==1.50.2
python-dotenv==1.0.1